pyproject.toml
[project]
name = "nativo"
version = "0.1.0"
description = "A platform for preserving endangered languages through digital archival of documents, words, and audio recordings"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=24.1.0",
    "black>=25.9.0",
    "email-validator>=2.3.0",
    "starlette-admin>=0.14.1",
    "itsdangerous>=2.2.0",
    "psycopg2-binary>=2.9.9",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.2",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
